WIP: Added stamina manager

+++ Assets/Scripts/Characters/CharacterData.cs
@@ -6,9 +6,9 @@ namespace Shipov_FP_Adventure
     public sealed class CharacterData : ScriptableObject
     {
         public string Name;
-        public int Stamina;
-        public int Health;
-        public int MaxHealth;
-        public int MaxStamina;
+        public float Stamina;
+        public float Health;
+        public float MaxHealth;
+        public float MaxStamina;
     }
 }
